Codec cleanup
