+++ src/config/codec.rs
@@ -1,0 +1,12 @@
+use super::*;
+
+impl Codec for ProtocolVersion {
+    fn encode(&self, buffer: &mut Vec<u8>) -> Result<(), CodecError> {
+        (*self as u8).encode(buffer)?;
+        Ok(())
+    }
+
+    fn decode(cursor: &mut Cursor) -> Result<Self, CodecError> {
+        Ok(Self::from(u8::decode(cursor)?)?)
+    }
+}
