+++ src/framing/codec.rs
@@ -257,3 +257,32 @@ impl<'a> Codec for MLSPlaintextCommitAuthData<'a> {
         Ok(())
     }
 }
+
+impl Codec for SenderType {
+    fn encode(&self, buffer: &mut Vec<u8>) -> Result<(), CodecError> {
+        (*self as u8).encode(buffer)?;
+        Ok(())
+    }
+    fn decode(cursor: &mut Cursor) -> Result<Self, CodecError> {
+        match SenderType::try_from(u8::decode(cursor)?) {
+            Ok(sender_type) => Ok(sender_type),
+            Err(_) => Err(CodecError::DecodingError),
+        }
+    }
+}
+
+impl Codec for Sender {
+    fn encode(&self, buffer: &mut Vec<u8>) -> Result<(), CodecError> {
+        self.sender_type.encode(buffer)?;
+        self.sender.encode(buffer)?;
+        Ok(())
+    }
+    fn decode(cursor: &mut Cursor) -> Result<Self, CodecError> {
+        let sender_type = SenderType::decode(cursor)?;
+        let sender = LeafIndex::from(u32::decode(cursor)?);
+        Ok(Sender {
+            sender_type,
+            sender,
+        })
+    }
+}
