+++ src/group/codec.rs
@@ -1,0 +1,33 @@
+use super::*;
+
+impl Codec for GroupId {
+    fn encode(&self, buffer: &mut Vec<u8>) -> Result<(), CodecError> {
+        encode_vec(VecSize::VecU8, buffer, &self.value)?;
+        Ok(())
+    }
+    fn decode(cursor: &mut Cursor) -> Result<Self, CodecError> {
+        let value = decode_vec(VecSize::VecU8, cursor)?;
+        Ok(GroupId { value })
+    }
+}
+
+impl Codec for GroupEpoch {
+    fn encode(&self, buffer: &mut Vec<u8>) -> Result<(), CodecError> {
+        self.0.encode(buffer)?;
+        Ok(())
+    }
+    fn decode(cursor: &mut Cursor) -> Result<Self, CodecError> {
+        let inner = u64::decode(cursor)?;
+        Ok(GroupEpoch(inner))
+    }
+}
+
+impl Codec for GroupContext {
+    fn encode(&self, buffer: &mut Vec<u8>) -> Result<(), CodecError> {
+        self.group_id.encode(buffer)?;
+        self.epoch.encode(buffer)?;
+        encode_vec(VecSize::VecU8, buffer, &self.tree_hash)?;
+        encode_vec(VecSize::VecU8, buffer, &self.confirmed_transcript_hash)?;
+        Ok(())
+    }
+}
