+++ src/tree/codec.rs
@@ -1,6 +1,8 @@
 use crate::tree::{node::*, secret_tree::*, *};
 use std::convert::TryFrom;
 
+// Nodes
+
 impl Codec for NodeType {
     fn encode(&self, buffer: &mut Vec<u8>) -> Result<(), CodecError> {
         (*self as u8).encode(buffer)?;
@@ -29,6 +31,25 @@ impl Codec for Node {
             node_type,
             key_package,
             node,
+        })
+    }
+}
+
+impl Codec for ParentNode {
+    fn encode(&self, buffer: &mut Vec<u8>) -> Result<(), CodecError> {
+        self.public_key.encode(buffer)?;
+        encode_vec(VecSize::VecU32, buffer, &self.unmerged_leaves)?;
+        encode_vec(VecSize::VecU8, buffer, &self.parent_hash)?;
+        Ok(())
+    }
+    fn decode(cursor: &mut Cursor) -> Result<Self, CodecError> {
+        let public_key = HPKEPublicKey::decode(cursor)?;
+        let unmerged_leaves = decode_vec(VecSize::VecU32, cursor)?;
+        let parent_hash = decode_vec(VecSize::VecU8, cursor)?;
+        Ok(ParentNode {
+            public_key,
+            unmerged_leaves,
+            parent_hash,
         })
     }
 }
@@ -70,6 +91,44 @@ impl Codec for UpdatePath {
 impl Codec for SecretTreeNode {
     fn encode(&self, buffer: &mut Vec<u8>) -> Result<(), CodecError> {
         self.secret.encode(buffer)?;
+        Ok(())
+    }
+}
+
+// Hash inputs
+
+impl<'a> Codec for ParentNodeHashInput<'a> {
+    fn encode(&self, buffer: &mut Vec<u8>) -> Result<(), CodecError> {
+        self.node_index.encode(buffer)?;
+        self.parent_node.encode(buffer)?;
+        encode_vec(VecSize::VecU8, buffer, &self.left_hash)?;
+        encode_vec(VecSize::VecU8, buffer, &self.right_hash)?;
+        Ok(())
+    }
+}
+
+impl<'a> Codec for LeafNodeHashInput<'a> {
+    fn encode(&self, buffer: &mut Vec<u8>) -> Result<(), CodecError> {
+        self.node_index.as_u32().encode(buffer)?;
+        self.key_package.encode(buffer)?;
+        Ok(())
+    }
+}
+
+// Index
+
+impl Codec for LeafIndex {
+    fn encode(&self, buffer: &mut Vec<u8>) -> Result<(), CodecError> {
+        self.0.encode(buffer)
+    }
+}
+
+// Secret tree
+
+impl Codec for TreeContext {
+    fn encode(&self, buffer: &mut Vec<u8>) -> Result<(), CodecError> {
+        self.node.encode(buffer)?;
+        self.generation.encode(buffer)?;
         Ok(())
     }
 }
