+++ src/tree/hash_input.rs
@@ -31,14 +31,14 @@
 use super::index::NodeIndex;
 use super::node::ParentNode;
 use crate::ciphersuite::Ciphersuite;
-use crate::codec::{encode_vec, Codec, CodecError, VecSize};
+use crate::codec::Codec;
 use crate::key_packages::KeyPackage;
 
 pub struct ParentNodeHashInput<'a> {
-    node_index: u32,
-    parent_node: &'a Option<ParentNode>,
-    left_hash: &'a [u8],
-    right_hash: &'a [u8],
+    pub(crate) node_index: u32,
+    pub(crate) parent_node: &'a Option<ParentNode>,
+    pub(crate) left_hash: &'a [u8],
+    pub(crate) right_hash: &'a [u8],
 }
 
 impl<'a> ParentNodeHashInput<'a> {
@@ -62,8 +62,8 @@ impl<'a> ParentNodeHashInput<'a> {
 }
 
 pub struct LeafNodeHashInput<'a> {
-    node_index: &'a NodeIndex,
-    key_package: &'a Option<KeyPackage>,
+    pub(crate) node_index: &'a NodeIndex,
+    pub(crate) key_package: &'a Option<KeyPackage>,
 }
 
 impl<'a> LeafNodeHashInput<'a> {
@@ -76,23 +76,5 @@ impl<'a> LeafNodeHashInput<'a> {
     pub fn hash(&self, ciphersuite: &Ciphersuite) -> Vec<u8> {
         let payload = self.encode_detached().unwrap();
         ciphersuite.hash(&payload)
-    }
-}
-
-impl<'a> Codec for ParentNodeHashInput<'a> {
-    fn encode(&self, buffer: &mut Vec<u8>) -> Result<(), CodecError> {
-        self.node_index.encode(buffer)?;
-        self.parent_node.encode(buffer)?;
-        encode_vec(VecSize::VecU8, buffer, &self.left_hash)?;
-        encode_vec(VecSize::VecU8, buffer, &self.right_hash)?;
-        Ok(())
-    }
-}
-
-impl<'a> Codec for LeafNodeHashInput<'a> {
-    fn encode(&self, buffer: &mut Vec<u8>) -> Result<(), CodecError> {
-        self.node_index.as_u32().encode(buffer)?;
-        self.key_package.encode(buffer)?;
-        Ok(())
     }
 }
